// src/shared/dataprovider/api/model/reportAccountFlowSummary.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Pagination } from './pagination';
import { ReportAccountFlowSummaryData } from './reportAccountFlowSummaryData';

export interface ReportAccountFlowSummary { 
    Pagination?: Pagination;
    Data?: Array<ReportAccountFlowSummaryData>;
}